Update convert-to-webm.sh
Previously metadata title would always be set as
whatever the output of the file would be.
Now it checks if the file already has a metadata
title and sets that as the metadata title.
If the file doesn't have a metadata title, it sets
metadata title as it previously did.