Add caching to ncmpcpp-art
Scripts use $XDG_CACHE_HOME/ncmpcpp instead of /tmp
If the directory doesn't exist ncmpcpp-art creates one

notify.sh checks if a cover with current album artist and
album exists. If not, it generates a 500x500 jpg cover for
the current song. Other songs by the same album artist in
the same album use the same cover. The only downside is,
if different songs have different cover art in the same
album only the first generated cover will be used.